requirements.txt
pandas 
numpy
pymongo[srv]
python-dotenv

-e .